bot/pyproject.toml
[project]
name = "forge-bot"
version = "0.1.0"
description = "Standalone Telegram bot for the forge development pipeline"
requires-python = ">=3.11"
dependencies = [
    "python-telegram-bot>=21.0",
    "anthropic>=0.40.0",
    "httpx>=0.28.1",
    "fastapi>=0.115.0",
    "uvicorn>=0.34.0",
]

[tool.hatch.build.targets.wheel]
packages = ["."]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "ruff>=0.15.7",
]
